chore: bump workspace to 0.4.1
Patch release rolling up the work since 0.4.0:
- ferrotorch-llama: Llama 3 70B GPU example, GGUF tensor-name remap,
  JSON-schema constrained-decoding logits processor (grammar module)
- ferrotorch-cubecl: quant kernel module, half/approx deps
- ferrotorch-llama: ProSparse 7B GPU example, serde_json wiring
- Cross-crate touch-ups across core/nn/optim/gpu/jit/serialize/vision/distributions
  carried over from the recent paged + ProSparse work

ferrotorch-paged remains gitignored (work-in-progress crate).

+++ ferrotorch-optim/src/scheduler/lambda_lr.rs
@@ -136,7 +136,7 @@ mod tests {
 
         for step in 1..=10 {
             sched.step(&mut opt);
-            let expected = 0.1 * gamma.powi(step as i32);
+            let expected = 0.1 * gamma.powi(step);
             assert!(
                 (opt.lr - expected).abs() < 1e-12,
                 "step {step}: expected {expected}, got {}",
